backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
sentence-transformers>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
ortools>=9.8.0
python-dotenv==1.0.0
bcrypt==4.1.1